Rewrite `stepParser` to remove the `ByteString` argument
The initial `ByteString` was used as the initial value for the parser's current
line of input. This led to problems with incremental parsing:

    > starve (feed "foo" (stepParser (                  many anyChar) mempty "bar"))
    Success "bar"

    > starve (feed "foo" (stepParser (release mempty *> many anyChar) mempty "bar"))
    Success "foo"

The initial chunk of input *should* be obtained from the input itself, using
`rewindIt` (that's basically what `release mempty` in the second example does).

This is a breaking change, since we remove an argument from the `stepParser`
function. However, client code usually relies on the more high-level
`parseString`/`parseByteString`, so this should not have too much impact.

Fixes #21.